Camera: fix deadlock when stopping acquisition.
FlyCapture2::Camera::RetrieveBuffer() will block until an image is
available or until the acquisition is stopped. If external triggering
is enabled and there are no trigger pulses, the acquisition can't be
stopped since stopAcq() will wait for the acquisition thread to finnish
and call _stopAcq()

Let stopAcq() stop the acquisition. FlyCapture2::Camera::RetrieveBuffer()
will return an error code that can be acted upon in the acquisition thread.